+++ src/builtins/env.c
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 19:28:26 by emadriga          #+#    #+#             */
-/*   Updated: 2021/12/24 16:19:10 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/01/23 08:20:40 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -112,10 +112,9 @@ void	init_ms_env(char **env_vector, t_str **env_list)
 /**
  * * This should recreate the bash funtion "env".
  * * Returms enviroment variables list
- * @param env_list	enviroment list
  * @param argv		vector of arguments (ERROR if not null)
 */
-void	ft_env(t_str **env_list, char **argv)
+void	ft_env(char **argv)
 {
 	t_str	*aux;
 
@@ -124,7 +123,7 @@ void	ft_env(t_str **env_list, char **argv)
 		log_error(ENV_NOT_OPTIONS_OR_ARGUMENT_ALLOWED, 1);
 	else
 	{
-		aux = *env_list;
+		aux = g_var.env;
 		while (aux != NULL)
 		{
 			if (ft_strchr(aux->str, '='))